[prog] Modularize getplayerdata route, edit api/data response
