+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "schema": "psql -U ballislife -d ballislife -f database/schema.sql",
-    "seed": "psql -U ballislife -d ballislife -f database/seed.sql"
+    "seed": "psql -U ballislife -d ballislife -f database/seed.sql",
+    "start": "nodemon server/index.js"
   },
   "repository": {
     "type": "git",
@@ -24,6 +25,7 @@
     "body-parser": "^1.18.3",
     "express": "^4.16.4",
     "morgan": "^1.9.1",
+    "nodemon": "^1.18.9",
     "pg": "^7.8.0"
   }
 }
