+++ server/playerData/controller.js
@@ -1,10 +1,10 @@
-const { getAllPlayersData } = require('../database/helpers');
+const { getAllPlayersData } = require('../../database/helpers');
 
 // CONTROLLER FUNCTIONS HERE - MAKE SURE EVERYONE REBASES
 module.exports = {
   getPlayerData: (req, res) => {
-    getAllPlayersData
-      .then((data) => { res.send(data); })
+    getAllPlayersData()
+      .then(({ rows }) => { res.send(rows); })
       .catch((err) => { console.error(err); });
   },
 };
