+++ server/router.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const { getPlayerData } = require('./controller');
+const { getPlayerData } = require('./playerData/controller');
 
 // ROUTES HERE - MAKE SURE EVERYONE REBASES
 router.route('/data/')
